refactor: moved files into similar folders as Jess to avoid duplication
