+++ lib/features/chirp/data/data.dart
@@ -3,8 +3,8 @@ export 'models/conversations/message_model.dart';
 export 'models/conversations/conversation_model_helper.dart';
 export 'models/conversations/message_model_helper.dart';
 
-export 'repository/conversations/conversation_repository_impl.dart';
-export 'repository/conversations/message_repository_impl.dart';
+export 'repositories/conversations/conversation_repository_impl.dart';
+export 'repositories/conversations/message_repository_impl.dart';
 
 export 'datasources/conversations/messaging_local_datasource.dart';
 export 'datasources/conversations/messaging_remote_datasource.dart';
