+++ lib/features/chirp/domain/domain.dart
@@ -1,7 +1,7 @@
 export 'entities/conversations/conversation.dart';
 export 'entities/conversations/message.dart';
-export 'repository/conversations/conversation_repository.dart';
-export 'repository/conversations/message_repository.dart';
+export 'repositories/conversations/conversation_repository.dart';
+export 'repositories/conversations/message_repository.dart';
 export 'usecases/conversations/get_conversations.dart';
 export 'usecases/conversations/get_messages.dart';
 export 'usecases/conversations/send_message.dart';
