+++ lib/features/chirp/domain/usecases/conversations/get_conversations.dart
@@ -1,7 +1,7 @@
 import 'package:academia/core/core.dart';
 import 'package:dartz/dartz.dart';
 import '../../entities/conversations/conversation.dart';
-import '../../repository/conversations/conversation_repository.dart';
+import '../../repositories/conversations/conversation_repository.dart';
 
 class GetConversations implements UseCase<List<Conversation>, NoParams> {
   final ConversationRepository repository;
