+++ lib/features/chirp/presentation/views/feed_screen.dart
@@ -2,8 +2,6 @@ import 'package:academia/features/chirp/presentation/bloc/feed/feed_bloc.dart';
 import 'package:academia/features/chirp/presentation/widgets/post_card.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:material_symbols_icons/material_symbols_icons.dart';
-import 'package:sliver_tools/sliver_tools.dart';
 
 class FeedPage extends StatelessWidget {
   const FeedPage({super.key});
