+++ lib/injection_container.dart
@@ -5,8 +5,8 @@ import 'package:academia/features/auth/auth.dart';
 import 'package:academia/features/auth/data/data.dart';
 import 'package:academia/features/chirp/data/datasources/conversations/messaging_local_datasource.dart';
 import 'package:academia/features/chirp/data/datasources/conversations/messaging_remote_datasource.dart';
-import 'package:academia/features/chirp/data/repository/conversations/conversation_repository_impl.dart';
-import 'package:academia/features/chirp/data/repository/conversations/message_repository_impl.dart';
+import 'package:academia/features/chirp/data/repositories/conversations/conversation_repository_impl.dart';
+import 'package:academia/features/chirp/data/repositories/conversations/message_repository_impl.dart';
 import 'package:academia/features/chirp/domain/usecases/conversations/get_conversations.dart';
 import 'package:academia/features/chirp/domain/usecases/conversations/get_messages.dart';
 import 'package:academia/features/chirp/domain/usecases/conversations/send_message.dart';
